+++ Source/Headers/bg3d.h
@@ -11,7 +11,7 @@
 
 #define	MAX_BG3D_MATERIALS			400			// max # of materials in a bg3d file
 
-#define	MAX_BG3D_GROUPS			(MODEL_GROUP_SKELETONBASE+MAX_SKELETON_TYPES)	// skeletons are @ end of list, so can use these counts for max #
+#define	MAX_BG3D_GROUPS			((int)MODEL_GROUP_SKELETONBASE+(int)MAX_SKELETON_TYPES)	// skeletons are @ end of list, so can use these counts for max #
 #define	MAX_OBJECTS_IN_GROUP	100
 
 		/***********************/
@@ -104,7 +104,7 @@ enum
 typedef struct
 {
 	uint32_t	type;								// geometry type
-	long		numMaterials;						// # material layers
+	int32_t		numMaterials;						// # material layers
 	uint32_t	layerMaterialNum[MAX_MULTITEXTURE_LAYERS];	// index into material list
 	uint32_t	flags;								// flags
 	uint32_t	numPoints;							// (if applicable)
